@@ -1,8 +1,8 @@
-import { injectable, inject, Guard } from '@micro-fleet/common';
+import { injectable, inject, Guard } from '@micro-fleet/common'
 
-import { Types } from '../Types';
-import { IDatabaseConnector } from '../connector/IDatabaseConnector';
-import { AtomicSessionFlow } from './AtomicSessionFlow';
+import { Types } from '../Types'
+import { IDatabaseConnector } from '../connector/IDatabaseConnector'
+import { AtomicSessionFlow } from './AtomicSessionFlow'
 
 
 /**
@@ -11,17 +11,17 @@ import { AtomicSessionFlow } from './AtomicSessionFlow';
 @injectable()
 export class AtomicSessionFactory {
 
-	constructor(
-		@inject(Types.DB_CONNECTOR) protected _dbConnector: IDatabaseConnector
-	) {
-		Guard.assertArgDefined('_dbConnector', _dbConnector);
-	}
+    constructor(
+        @inject(Types.DB_CONNECTOR) protected _dbConnector: IDatabaseConnector
+    ) {
+        Guard.assertArgDefined('_dbConnector', _dbConnector)
+    }
 
-	/**
-	 * Starts executing queries in transactions.
-	 * @param {string[]} names Only executes the queries on connections with specified names.
-	 */
-	public startSession(): AtomicSessionFlow {
-		return new AtomicSessionFlow(this._dbConnector);
-	}
+    /**
+     * Starts executing queries in transactions.
+     * @param {string[]} names Only executes the queries on connections with specified names.
+     */
+    public startSession(): AtomicSessionFlow {
+        return new AtomicSessionFlow(this._dbConnector)
+    }
 }
