@@ -1,5 +1,5 @@
 export class Types {
-	public static readonly DB_ADDON = 'persistence.DatabaseAddOn';
-	public static readonly DB_CONNECTOR = 'persistence.IDatabaseConnector';
-	public static readonly ATOMIC_SESSION_FACTORY = 'persistence.AtomicSessionFactory';
+    public static readonly DB_ADDON = 'persistence.DatabaseAddOn'
+    public static readonly DB_CONNECTOR = 'persistence.IDatabaseConnector'
+    public static readonly ATOMIC_SESSION_FACTORY = 'persistence.AtomicSessionFactory'
 }
