@@ -1,6 +1,6 @@
-import { PagedArray } from '@micro-fleet/common';
+import { PagedArray } from '@micro-fleet/common'
 
-import { AtomicSession } from './atom/AtomicSession';
+import { AtomicSession } from './atom/AtomicSession'
 
 
 /**
@@ -9,28 +9,28 @@ import { AtomicSession } from './atom/AtomicSession';
  * @deprecated
  */
 export interface RepositoryOptions {
-	/**
-	 * A transaction to which this operation is restricted.
-	 */
-	atomicSession?: AtomicSession;
+    /**
+     * A transaction to which this operation is restricted.
+     */
+    atomicSession?: AtomicSession
 
-	/**
-	 * Account ID.
-	 */
-	accountId?: BigInt;
+    /**
+     * Account ID.
+     */
+    accountId?: bigint
 }
 
 export interface RepositoryExistsOptions extends RepositoryOptions {
-	/**
-	 * Whether to exclude records marked as soft-deleted.
-	 * Default to `false`.
-	 */
-	excludeDeleted?: boolean;
+    /**
+     * Whether to exclude records marked as soft-deleted.
+     * Default to `false`.
+     */
+    excludeDeleted?: boolean
 
-	/**
-	 * Tenant ID.
-	 */
-	tenantId?: BigInt;
+    /**
+     * Tenant ID.
+     */
+    tenantId?: bigint
 }
 
 export interface RepositoryCountAllOptions extends RepositoryExistsOptions {
@@ -43,12 +43,12 @@ export interface RepositoryDeleteOptions extends RepositoryOptions {
 }
 
 export interface RepositoryFindOptions extends RepositoryOptions {
-	version?: number;
+    version?: number
 }
 
 export interface RepositoryPageOptions extends RepositoryCountAllOptions {
-	sortBy?: string;
-	sortType?: string;
+    sortBy?: string
+    sortType?: string
 }
 
 export interface RepositoryPatchOptions extends RepositoryOptions {
@@ -64,7 +64,7 @@ export interface RepositorySetMainOptions extends RepositoryOptions {
 }
 
 export interface RepositoryDelVersionOptions extends RepositoryOptions {
-	olderThan?: Date;
+    olderThan?: Date
 }
 
 export interface RepositoryRestrictOptions extends RepositoryOptions {
@@ -73,108 +73,108 @@ export interface RepositoryRestrictOptions extends RepositoryOptions {
 /**
  * Provides common CRUD operations, based on Unit of Work pattern.
  */
-export interface IRepository<TModel extends IModelDTO, TPk extends PkType = BigInt, TUk = NameUk> {
+export interface IRepository<TModel extends IModelDTO, TPk extends PkType = bigint, TUk = NameUk> {
 
-	/**
-	 * Counts all records in a table.
-	 */
-	countAll(options?: RepositoryCountAllOptions): Promise<number>;
+    /**
+     * Counts all records in a table.
+     */
+    countAll(options?: RepositoryCountAllOptions): Promise<number>
 
-	/**
-	 * Inserts one or more `model` to database.
-	 * @param {DTO model} model The model to be inserted.
-	 */
-	create(model: TModel | TModel[], options?: RepositoryCreateOptions): Promise<TModel & TModel[]>;
+    /**
+     * Inserts one or more `model` to database.
+     * @param {DTO model} model The model to be inserted.
+     */
+    create(model: TModel | TModel[], options?: RepositoryCreateOptions): Promise<TModel | TModel[]>
 
-	/**
-	 * Permanently deletes one or many records.
-	 * @param {PK Type} pk The primary key object.
-	 */
-	deleteHard(pk: TPk | TPk[], options?: RepositoryDeleteOptions): Promise<number>;
+    /**
+     * Permanently deletes one or many records.
+     * @param {PK Type} pk The primary key object.
+     */
+    deleteHard(pk: TPk | TPk[], options?: RepositoryDeleteOptions): Promise<number>
 
-	/**
-	 * Checks if a record exists or not.
-	 * @param {TUk} props An object with non-primary unique properties.
-	 */
-	exists(props: TUk, options?: RepositoryExistsOptions): Promise<boolean>;
-	
-	/**
-	 * Selects only one record with `pk`.
-	 * @param {PK Type} pk The primary key object.
-	 */
-	findByPk(pk: TPk, options?: RepositoryFindOptions): Promise<TModel>;
+    /**
+     * Checks if a record exists or not.
+     * @param {TUk} props An object with non-primary unique properties.
+     */
+    exists(props: TUk, options?: RepositoryExistsOptions): Promise<boolean>
 
-	/**
-	 * Selects `pageSize` number of records at page `pageIndex`.
-	 * @param {number} pageIndex Index of the page.
-	 * @param {number} pageSize Number of records in a page.
-	 */
-	page(pageIndex: number, pageSize: number, options?: RepositoryPageOptions): Promise<PagedArray<TModel>>;
+    /**
+     * Selects only one record with `pk`.
+     * @param {PK Type} pk The primary key object.
+     */
+    findByPk(pk: TPk, options?: RepositoryFindOptions): Promise<TModel>
 
-	/**
-	 * Updates new value for specified properties in `model`.
-	 */
-	patch(model: Partial<TModel> | Partial<TModel>[], options?: RepositoryPatchOptions): Promise<Partial<TModel> & Partial<TModel>[]>;
+    /**
+     * Selects `pageSize` number of records at page `pageIndex`.
+     * @param {number} pageIndex Index of the page.
+     * @param {number} pageSize Number of records in a page.
+     */
+    page(pageIndex: number, pageSize: number, options?: RepositoryPageOptions): Promise<PagedArray<TModel>>
 
-	/**
-	 * Replaces a record with `model`.
-	 */
-	update(model: TModel | TModel[], options?: RepositoryUpdateOptions): Promise<TModel & TModel[]>;
+    /**
+     * Updates new value for specified properties in `model`.
+     */
+    patch(model: Partial<TModel> | Partial<TModel>[], options?: RepositoryPatchOptions): Promise<Partial<TModel> | Partial<TModel>[]>
+
+    /**
+     * Replaces a record with `model`.
+     */
+    update(model: TModel | TModel[], options?: RepositoryUpdateOptions): Promise<TModel | TModel[]>
 }
 
 /**
  * Provides common operations to soft-delete and recover models.
  */
-export interface ISoftDelRepository<TModel extends IModelDTO, TPk extends PkType = BigInt, TUk = NameUk>
-		extends IRepository<TModel, TPk, TUk> {
+export interface ISoftDelRepository<TModel extends IModelDTO, TPk extends PkType = bigint, TUk = NameUk>
+        extends IRepository<TModel, TPk, TUk> {
 
-	/**
-	 * Marks one or many records with `pk` as deleted.
-	 * @param {PK Type} pk The primary key object.
-	 */
-	deleteSoft(pk: TPk | TPk[], options?: RepositoryDeleteOptions): Promise<number>;
+    /**
+     * Marks one or many records with `pk` as deleted.
+     * @param {PK Type} pk The primary key object.
+     */
+    deleteSoft(pk: TPk | TPk[], options?: RepositoryDeleteOptions): Promise<number>
 
-	/**
-	 * Marks one or many records with `pk` as NOT deleted.
-	 * @param {PK Type} pk The primary key object.
-	 */
-	recover(pk: TPk | TPk[], options?: RepositoryRecoverOptions): Promise<number>;
+    /**
+     * Marks one or many records with `pk` as NOT deleted.
+     * @param {PK Type} pk The primary key object.
+     */
+    recover(pk: TPk | TPk[], options?: RepositoryRecoverOptions): Promise<number>
 
 }
 
 /**
  * Provides common operations to control models' revisions.
  */
-export interface IVersionRepository<TModel extends IVersionControlled, TPk extends PkType = BigInt, TUk = NameUk>
-		extends ISoftDelRepository<TModel, TPk, TUk> {
+export interface IVersionRepository<TModel extends IVersionControlled, TPk extends PkType = bigint, TUk = NameUk>
+        extends ISoftDelRepository<TModel, TPk, TUk> {
 
-	/**
-	 * Permanently deletes one or many version of a record.
-	 * Can be filtered with `olderThan` option.
-	 * @param {PK Type} pk The primary key object.
-	 */
-	deleteHardVersions(pk: TPk, versions: number | number[], options?: RepositoryDelVersionOptions): Promise<number>;
+    /**
+     * Permanently deletes one or many version of a record.
+     * Can be filtered with `olderThan` option.
+     * @param {PK Type} pk The primary key object.
+     */
+    deleteHardVersions(pk: TPk, versions: number | number[], options?: RepositoryDelVersionOptions): Promise<number>
 
-	/**
-	 * Selects `pageSize` number of version of a record at page `pageIndex`.
-	 * @param {PK Type} pk The primary key object.
-	 * @param {number} pageIndex Index of the page.
-	 * @param {number} pageSize Number of records in a page.
-	 */
-	pageVersions(pk: TPk, pageIndex: number, pageSize: number, options?: RepositoryPageOptions): Promise<number>;
+    /**
+     * Selects `pageSize` number of version of a record at page `pageIndex`.
+     * @param {PK Type} pk The primary key object.
+     * @param {number} pageIndex Index of the page.
+     * @param {number} pageSize Number of records in a page.
+     */
+    pageVersions(pk: TPk, pageIndex: number, pageSize: number, options?: RepositoryPageOptions): Promise<number>
 
-	/**
-	 * Marks a revision as main version of the record with `pk`.
-	 * @param {PK Type} pk The primary key object.
-	 * @param {number} version The version number.
-	 */
-	setAsMain(pk: TPk, version: number, options?: RepositorySetMainOptions): Promise<number>;
+    /**
+     * Marks a revision as main version of the record with `pk`.
+     * @param {PK Type} pk The primary key object.
+     * @param {number} version The version number.
+     */
+    setAsMain(pk: TPk, version: number, options?: RepositorySetMainOptions): Promise<number>
 
-	/**
-	 * Removes old versions to keep number of version to be equal or less than `nVersion`.
-	 * @param {PK Type} pk The primary key object.
-	 * @param {number} nVersion Number of versions to keep.
-	 */
-	
-	restrictQuantity(pk: TPk, nVersion: number, options?: RepositoryRestrictOptions): void;
+    /**
+     * Removes old versions to keep number of version to be equal or less than `nVersion`.
+     * @param {PK Type} pk The primary key object.
+     * @param {number} nVersion Number of versions to keep.
+     */
+
+    restrictQuantity(pk: TPk, nVersion: number, options?: RepositoryRestrictOptions): void
 }
